FIR IDE: Extract common logic from KotlinTypeArgumentInfoHandlerBase to a base class.
